Create Chinese (Traditional)
The only way to create a folder for Chinese (Traditional) is through the Desktop Github Software which for some reason does bot permit me to create PR.  So, I can only "propose new file" here.

There are also some missing strings in this English PO file, namely, in the control panel the "Navigation" button has no strings.  The Place Search of Foursquare and Eventful strings are missing too.  When these missing strings in the English PO file are made good, I'll be more than happy to do the translation again.